submissions-team/marinath-jeevanantham/requirements.txt
# Contains all libraries that are needed for this project
streamlit
tensorflow
opencv-python
pillow
numpy
matplotlib
huggingface_hub